Move sum-all-objects-numbers util into separate file
