+++ workers/loc.api/helpers/utils.js
@@ -185,32 +185,6 @@ const pickAllLowerObjectsNumbers = (propName, objects = []) => {
   }, {})
 }
 
-const sumAllObjectsNumbers = (propName, objects = []) => {
-  return objects.reduce((accum, curr) => {
-    if (
-      !curr?.[propName] ||
-      typeof curr?.[propName] !== 'object'
-    ) {
-      return accum
-    }
-
-    const entries = Object.entries(curr[propName])
-
-    return entries.reduce((accum, [key, val]) => {
-      const prevVal = Number.isFinite(accum?.[key])
-        ? accum[key]
-        : 0
-      const currVal = Number.isFinite(val)
-        ? val
-        : 0
-
-      accum[key] = prevVal + currVal
-
-      return accum
-    }, accum)
-  }, {})
-}
-
 module.exports = {
   checkParamsAuth,
   tryParseJSON,
@@ -219,6 +193,5 @@ module.exports = {
   isNotSyncRequired,
   sumObjectsNumbers,
   pickLowerObjectsNumbers,
-  sumAllObjectsNumbers,
   pickAllLowerObjectsNumbers
 }
