+++ workers/loc.api/helpers/utils/index.js
@@ -3,8 +3,10 @@
 const orderBy = require('./order-by')
 const pushLargeArr = require('./push-large-arr')
 const sumArrayVolumes = require('./sum-array-volumes')
+const sumAllObjectsNumbers = require('./sum-all-objects-numbers')
 
 module.exports = {
+  sumAllObjectsNumbers,
   sumArrayVolumes,
   pushLargeArr,
   orderBy
